refactor(configuration): centralize header merging
