+++ experimental/packages/configuration/src/FileConfigFactory.ts
@@ -4,7 +4,7 @@
  */
 
 import { diag } from '@opentelemetry/api';
-import { getStringFromEnv } from '@opentelemetry/core';
+import { getStringFromEnv, parseKeyPairsIntoRecord } from '@opentelemetry/core';
 import * as fs from 'fs';
 import * as yaml from 'yaml';
 import type { ConfigFactory } from './IConfigFactory';
@@ -12,6 +12,7 @@ import { substituteEnvVars } from './utils';
 import type {
   AttributeNameValue,
   ConfigurationModel,
+  NameStringValuePair,
   TextMapPropagator,
 } from './generated/types';
 // Pre-compiled ajv validator — eliminates runtime schema compilation on cold start.
@@ -170,6 +171,27 @@ export function mergeResourceAttributesConfig(
   }
 
   return mergedAttrs;
+}
+
+/**
+ * Merge headers from `headers` and `headers_list` (comma-separated key=value
+ * pairs) into a record. Entries in `headers` take precedence.
+ */
+export function mergeHeadersConfig(
+  headers?: NameStringValuePair[],
+  headers_list?: string | null
+): Record<string, string> | undefined {
+  if (!headers && !headers_list) {
+    return undefined;
+  }
+
+  const mergedHeaders = parseKeyPairsIntoRecord(headers_list ?? undefined);
+  headers?.forEach(header => {
+    if (header.value !== null) {
+      mergedHeaders[header.name] = header.value;
+    }
+  });
+  return mergedHeaders;
 }
 
 /**
