+++ experimental/packages/opentelemetry-sdk-node/test/utils.test.ts
@@ -9,8 +9,6 @@ import {
   getLoggerProviderConfigFromEnv,
   getBatchLogRecordProcessorConfigFromEnv,
   getResourceDetectorsFromConfiguration,
-  getGrpcMetadataFromHeaders,
-  getHeadersFromConfiguration,
 } from '../src/utils';
 import * as assert from 'assert';
 import * as sinon from 'sinon';
@@ -407,72 +405,5 @@ describe('getResourceDetectorsFromConfiguration', function () {
       processDetector,
       serviceInstanceIdDetector,
     ]);
-  });
-});
-
-describe('getHeadersFromConfiguration', function () {
-  it('returns empty object when headers are not set', function () {
-    const config: ConfigurationModel = {};
-    assert.deepStrictEqual(
-      getHeadersFromConfiguration(
-        config.tracer_provider?.processors?.[0]?.simple?.exporter?.otlp_http
-          ?.headers
-      ),
-      undefined
-    );
-  });
-
-  it('returns headers object when headers are set', function () {
-    const config: ConfigurationModel = {
-      tracer_provider: {
-        processors: [
-          {
-            simple: {
-              exporter: {
-                otlp_http: {
-                  headers: [{ name: 'x-test-header', value: 'test-value' }],
-                },
-              },
-            },
-          },
-        ],
-      },
-    };
-    assert.deepStrictEqual(
-      getHeadersFromConfiguration(
-        config.tracer_provider?.processors?.[0]?.simple?.exporter?.otlp_http
-          ?.headers
-      ),
-      { 'x-test-header': 'test-value' }
-    );
-  });
-
-  it('parses headers_list and lets headers take precedence', function () {
-    assert.deepStrictEqual(
-      getHeadersFromConfiguration(
-        [
-          { name: 'shared', value: 'from-headers' },
-          { name: 'ignored', value: null },
-        ],
-        'shared=from-list,list-only=hello%20world,ignored=from-list'
-      ),
-      {
-        shared: 'from-headers',
-        'list-only': 'hello world',
-        ignored: 'from-list',
-      }
-    );
-  });
-});
-
-describe('getGrpcMetadataFromHeaders', function () {
-  it('merges headers_list into metadata with headers taking precedence', function () {
-    const metadata = getGrpcMetadataFromHeaders(
-      [{ name: 'shared', value: 'from-headers' }],
-      'shared=from-list,list-only=value'
-    );
-
-    assert.deepStrictEqual(metadata?.get('shared'), ['from-headers']);
-    assert.deepStrictEqual(metadata?.get('list-only'), ['value']);
   });
 });
